# api/internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "milmil API",
    "version": "1.0.0",
    "description": "Self-hosted anime media server"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development"
    }
  ],
  "paths": {
...
    "/api/v1/libraries/{id}": {
      "get": {
        "summary": "Get a library by ID",
        "operationId": "getLibrary",
        "tags": [
          "Libraries"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Library details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Library"
                }
              }
            }
          },
          "404": {
            "description": "Library not found"
          }
        }
      },
      "put": {
        "summary": "Update a library",
        "operationId": "updateLibrary",
        "tags": [
          "Libraries"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "path"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "path": {
                    "type": "string"
                  },
                  "enabled": {
                    "type": "boolean"
                  },
                  "scan_interval_minutes": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Library updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Library"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Library not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a library",
        "operationId": "deleteLibrary",
        "tags": [
          "Libraries"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Library deleted"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API token (mlml_ prefix)",
        "description": "Pass an API token obtained from POST /api/v1/auth/login or POST /api/v1/api-tokens. Tokens have the prefix `mlml_`."
      }
    },
    "schemas": {
...
      "Library": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "enabled": {
            "type": "integer",
            "description": "1 = enabled, 0 = disabled"
          },
          "scan_interval_minutes": {
            "type": "integer"
          },
          "last_scanned_at": {
            "type": "string",
            "nullable": true
          },
          "created_at": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}